[~] Isolated the enums and datum classes.
