+++ Assets/HoloToolkit/UX/Scripts/Buttons/AnimatorControllerAction.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: a983b48aab3a3bf4ca0e54b541a5c29c
+timeCreated: 1507327871
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
