+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonDatums.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 427b1cb036928334ea2eb36b49a91546
+folderAsset: yes
+timeCreated: 1507327985
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
