+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonDatums/MeshButtonDatum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 5fed3e269857bda429e953648f5aa46a
+timeCreated: 1507327871
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
