+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonDatums/SpriteButtonDatum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: e7764c2519171734dae2f1cf281a2ea7
+timeCreated: 1507327871
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
