+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonEnums.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 227974e60590efc49a177c23523c5ab3
+folderAsset: yes
+timeCreated: 1507326959
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
