+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonEnums/ButtonStateEnum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: ef89da7015edd334b8c7c6fd0c4a177d
+timeCreated: 1507327205
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
