+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonEnums/ToggleBehaviorEnum.cs
@@ -1,0 +1,14 @@
+﻿//
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+//
+
+namespace HoloToolkit.Unity.Buttons
+{
+    public enum ToggleBehaviorEnum
+    {
+        Manual,     // Toggle manually by changing the State property
+        OnTapped,   // Toggle automatically when button is tapped
+        OnFocus,    // Toggle automatically when button focus is entered;
+    }
+}
