+++ Assets/HoloToolkit/UX/Scripts/Buttons/ButtonEnums/ToggleBehaviorEnum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 290b408286e58b94eb0ac9eee68d2a39
+timeCreated: 1507327871
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
