+++ Assets/HoloToolkit/UX/Scripts/Buttons/SpriteButton.cs
@@ -3,8 +3,6 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 //
 using UnityEngine;
-using System;
-using System.Collections;
 
 namespace HoloToolkit.Unity.Buttons
 {
@@ -14,36 +12,6 @@ namespace HoloToolkit.Unity.Buttons
     [RequireComponent(typeof(SpriteRenderer))]
     public class SpriteButton : Button
     {
-        /// <summary>
-        /// Sprite Button State Data Set
-        /// </summary>
-        [Serializable]
-        public class SpriteButtonDatum
-        {
-            public SpriteButtonDatum(ButtonStateEnum state) { this.ActiveState = state; this.Name = state.ToString(); }
-
-            /// <summary>
-            /// Name of Datum entry
-            /// </summary>
-            public string Name;
-            /// <summary>
-            /// Button State association
-            /// </summary>
-            public ButtonStateEnum ActiveState;
-            /// <summary>
-            /// Button sprite for new state
-            /// </summary>
-            public Sprite ButtonSprite;
-            /// <summary>
-            /// Color for sprite in new state
-            /// </summary>
-            public Color SpriteColor = Color.white;
-            /// <summary>
-            /// New scale for button state
-            /// </summary>
-            public Vector3 Scale = Vector3.one;
-        }
-
         /// <summary>
         /// Button State data set for different interaction states
         /// </summary>
