+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 46b5b95c7eb1b2249960cadfbab1771c
+folderAsset: yes
+timeCreated: 1507326833
+licenseType: Pro
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
