+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/LayoutTypeEnum.cs
@@ -1,0 +1,16 @@
+﻿//
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+//
+
+namespace HoloToolkit.Unity.Collections
+{
+    /// <summary>
+    /// Collection layout type enum
+    /// </summary>
+    public enum LayoutTypeEnum
+    {
+        ColumnThenRow,          // Sort by column, then by row
+        RowThenColumn,          // Sort by row, then by column
+    }
+}
