+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/LayoutTypeEnum.cs.meta
