+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/OrientTypeEnum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 5a420dd2e9f0b3f4cba4c0b7881e6654
+timeCreated: 1507326833
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
