+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/SortTypeEnum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: a1daa1fc95763f84f8b0fe0ca0ee09eb
+timeCreated: 1507326833
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
