+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/SurfaceTypeEnum.cs
@@ -1,0 +1,18 @@
+﻿//
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License. See LICENSE in the project root for license information.
+//
+
+namespace HoloToolkit.Unity.Collections
+{
+    /// <summary>
+    /// The type of surface to map the collect to.
+    /// </summary>
+    public enum SurfaceTypeEnum
+    {
+        Cylinder,
+        Plane,
+        Sphere,
+        Scatter,
+    }
+}
