+++ Assets/HoloToolkit/UX/Scripts/Collections/CollectionEnums/SurfaceTypeEnum.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: c6f08a4a536b0e941acc54d44b1082f6
+timeCreated: 1507326833
+licenseType: Pro
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
